requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pydub==0.25.1
python-dotenv==1.0.1
pytubefix==8.12.1
redis==5.2.1
requests==2.32.3
spotipy==2.24.0
Unidecode==1.3.8
urllib3==2.3.0
